requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-crispy-forms==1.14.0
Pillow==9.2.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
djangorestframework==3.13.1